vimrc: Change working directory with netrw
Change working directory to be same as browsing directory.
As much as I would love to have the functionality be the same as my
function that finds the project directory (if it is a git project) and
sets that to the working directory, netrw does not seem to be easy to
hook into. It does not play nicely with autocommands like BufEnter etc.